feat: Schedule runners over multiple hosts
This is a very simple and naive implementation to start the service with
a pool of runners. A HostManager will then select the one which looks
the least busy based on an in-memory record of where it has put things.

This is fine for now, but if this project continues we will want a
smarter and more robust way of scheduling. Not to mention more
extendable. For example in the current design, if I wanted to extend the
Manager to be able to select based on runner tags (eg I wanted ARM), I
would probably have to re-write a good chunk.